fix: eliminate all float() from monetary/response code across codebase
Replace every float() call with money() (shared constant helper) or
Decimal-based arithmetic. Only 3 float() remain in the entire codebase:
- money() itself (the single conversion point for JSON serialization)
- 2 geographic coordinate env var parsers (non-monetary)

Files changed: 24 modules across fiscal, turns, hardware, inventory,
remote, sync, dashboard, enterprise_dashboard, and control-plane.

+++ backend/modules/dashboard/routes.py
@@ -266,10 +266,10 @@ async def get_ai_dashboard(auth: dict = Depends(verify_token), db=Depends(get_db
         "data": {
             "alerts": [{
                 "product_name": p["name"],
-                "urgency": "CRITICAL" if float(p.get("stock") or 0) <= 2 else "WARNING",
-                "current_stock": float(p.get("stock") or 0),
-                "days_until_stockout": max(1, int(float(p.get("stock") or 0) / max(float(p.get("sold_30d") or 0) / 30, 0.01))),
-                "recommended_order": int(float(p.get("min_stock") or 5) * 2),
+                "urgency": "CRITICAL" if money(p.get("stock") or 0) <= 2 else "WARNING",
+                "current_stock": money(p.get("stock") or 0),
+                "days_until_stockout": max(1, int(money(p.get("stock") or 0) / max(money(p.get("sold_30d") or 0) / 30, 0.01))),
+                "recommended_order": int(money(p.get("min_stock") or 5) * 2),
             } for p in low_stock],
             "top_products": [{
                 "name": t["name"],
